Remove input symbols
