+++ pinot-segment-local/src/test/java/org/apache/pinot/segment/local/utils/nativefst/mutablefst/MutableFSTImplTest.java
@@ -10,13 +10,13 @@ import static org.testng.AssertJUnit.assertEquals;
 import static org.testng.AssertJUnit.assertTrue;
 
 
-public class MutableFSTTest {
+public class MutableFSTImplTest {
   @Test
   public void shouldCompactNulls1()
       throws Exception {
     List<Integer> listGood = Lists.newArrayList(1, 2, 3, 4, 5, 6, 7, 8, 9);
     List<Integer> listBad = Lists.newArrayList(null, 1, 2, null, 3, 4, null, 5, 6, null, 7, 8, 9, null);
-    MutableFST.compactNulls((ArrayList) listBad);
+    MutableFSTImpl.compactNulls((ArrayList) listBad);
     assertEquals(listGood, listBad);
   }
 
@@ -24,14 +24,13 @@ public class MutableFSTTest {
   public void shouldCompactNulls2()
       throws Exception {
     ArrayList<Integer> listBad = (ArrayList) Lists.newArrayList(1);
-    MutableFST.compactNulls(listBad);
+    MutableFSTImpl.compactNulls(listBad);
     assertEquals(Lists.newArrayList(1), listBad);
   }
 
   @Test
-  public void shouldCreateWithStateSymbols()
-      throws Exception {
-    MutableFST fst = createStateSymbolFst();
+  public void shouldCreateWithStateSymbols() {
+    MutableFSTImpl fst = createStateSymbolFst();
 
     assertEquals(5, fst.getStateCount());
     assertEquals(1, fst.getState(0).getArcCount()); // start
@@ -40,14 +39,12 @@ public class MutableFSTTest {
     assertEquals(0, fst.getState(3).getArcCount()); // _C
     assertEquals(0, fst.getState(4).getArcCount()); // _D
 
-    assertEquals(4, fst.getInputSymbols().size());
     assertEquals(4, fst.getOutputSymbols().size());
 
     MutableState stateA = fst.getState(2);
     assertEquals("_A", fst.getStateSymbols().invert().keyForId(stateA.getId()));
     assertEquals(1, stateA.getArcCount());
     MutableArc arc = stateA.getArc(0);
-    assertEquals(fst.lookupInputSymbol("b"), arc.getIlabel());
     assertEquals(fst.lookupOutputSymbol("B"), arc.getOlabel());
     assertEquals(fst.getState("_B").getId(), arc.getNextState().getId());
     assertTrue(arc.hashCode() != 0);
@@ -56,22 +53,19 @@ public class MutableFSTTest {
 
   @Test
   public void shouldCopyWithTranslatedSymbols() {
-    MutableFST fst = new MutableFST();
+    MutableFSTImpl fst = new MutableFSTImpl();
     MutableState s0 = fst.newStartState("<start>");
     MutableState s1 = fst.newState();
     MutableState s2 = fst.newState();
-    fst.getInputSymbols().getOrAdd("a");
-    fst.getInputSymbols().getOrAdd("b");
-    fst.getInputSymbols().getOrAdd("c");
-    fst.getInputSymbols().getOrAdd("d");
+
     fst.getOutputSymbols().getOrAdd("A");
     fst.getOutputSymbols().getOrAdd("B");
     fst.getOutputSymbols().getOrAdd("C");
     fst.getOutputSymbols().getOrAdd("D");
-    fst.addArc(s0, "a", "A", s1);
-    fst.addArc(s1, "b", "B", s2);
-    fst.addArc(s0, "c", "C", s2);
-    fst.addArc(s2, "d", "D", s2);
+    fst.addArc(s0, "a", s1);
+    fst.addArc(s1, "b", s2);
+    fst.addArc(s0, "c", s2);
+    fst.addArc(s2, "d", s2);
 
     MutableSymbolTable newIn = new MutableSymbolTable();
     newIn.put("a", 101);
@@ -84,37 +78,61 @@ public class MutableFSTTest {
     newOut.put("C", 203);
     newOut.put("D", 204);
 
-    MutableFST result = MutableFST.copyFrom(fst);
+    MutableFSTImpl result = MutableFSTImpl.copyFrom(fst);
 
     MutableState rs0 = result.getState(0);
     MutableState rs1 = result.getState(1);
     MutableState rs2 = result.getState(2);
-    assertEquals(101, rs0.getArc(0).getIlabel());
+
     assertEquals(201, rs0.getArc(0).getOlabel());
-    assertEquals(103, rs0.getArc(1).getIlabel());
+
     assertEquals(203, rs0.getArc(1).getOlabel());
 
-    assertEquals(102, rs1.getArc(0).getIlabel());
+
     assertEquals(202, rs1.getArc(0).getOlabel());
 
-    assertEquals(104, rs2.getArc(0).getIlabel());
     assertEquals(204, rs2.getArc(0).getOlabel());
   }
 
-  private MutableFST createStateSymbolFst() {
-    MutableFST fst = new MutableFST();
+  private MutableFSTImpl createStateSymbolFst() {
+    MutableFSTImpl fst = new MutableFSTImpl();
 
-    MutableState startState = fst.newStartState("<start>");
+    fst.newStartState("<start>");
 
     // creating a few symbols by hand, others will get created automatically
     fst.newState("_B");
-    int inputA = fst.getInputSymbols().getOrAdd("a");
 
-    fst.addArc("<start>", "a", "A", "_A");
-    fst.addArc("_A", "b", "B", "_B");
-    fst.addArc("_B", "c", "C", "_C");
-    fst.addArc("_B", "d", "D", "_D");
+    fst.addArc("<start>", "a", "_A");
+    fst.addArc("_A", "b", "_B");
+    fst.addArc("_B", "c", "_C");
+    fst.addArc("_B", "d", "_D");
 
     return fst;
+  }
+
+  @Test
+  public void testTraversalFoo() {
+    MutableFSTImpl fst = createStateSymbolFst();
+
+    assertEquals(5, fst.getStateCount());
+    assertEquals(1, fst.getState(0).getArcCount()); // start
+    assertEquals(2, fst.getState(1).getArcCount()); // _B
+    assertEquals(1, fst.getState(2).getArcCount()); // _A
+    assertEquals(0, fst.getState(3).getArcCount()); // _C
+    assertEquals(0, fst.getState(4).getArcCount()); // _D
+
+    List<MutableArc> arcs = fst.getStartState().getArcs();
+
+    assertEquals(4, fst.getOutputSymbols().size());
+
+    MutableState stateA = fst.getState(2);
+    assertEquals("_A", fst.getStateSymbols().invert().keyForId(stateA.getId()));
+    assertEquals(1, stateA.getArcCount());
+    MutableArc arc = stateA.getArc(0);
+
+    assertEquals(fst.lookupOutputSymbol("B"), arc.getOlabel());
+    assertEquals(fst.getState("_B").getId(), arc.getNextState().getId());
+    assertTrue(arc.hashCode() != 0);
+    assertTrue(StringUtils.isNotBlank(arc.toString()));
   }
 }
